### gradle을 통한 모듈 분리 > - web,api,mongodb 로 도메인 별 모듈 분리
