+++ mongodb/init_data_mongo/vending-machine/drink.json
