+++ mongodb/src/main/java/com/example/drink/controller/MainController.java
@@ -1,0 +1,1 @@
+package com.example.drink.controller;
