+++ settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'drink'
+include 'api'
+include 'web'
+include 'mongodb'
+
