--- src/main/java/com/example/config/AppConfig.java
@@ -1,30 +1,0 @@
-package com.example.config;
-
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.mongo.MongoProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-@Configuration
-@EnableMongoRepositories(basePackages = {"com.example.drink.repo"})
-public class AppConfig {
-
-    private MongoProperties props = null;
-
-    @Value("${spring.data.mongodb.uri}")
-    private String mongoUrl;
-
-    @Bean
-    public MongoClient mongoClient() {
-        return MongoClients.create(mongoUrl);
-    }
-
-    @Bean
-    public MongoTemplate mongoTemplate() {
-        return new MongoTemplate(mongoClient(), "vending-machine");
-    }
-}
