--- src/test/resources/application-test.yml
@@ -1,17 +1,0 @@
-drink:
-  product:
-      list: 콜라,사이다,식혜
-spring:
-  data:
-    mongodb:
-      uri: mongodb://localhost:27017
-
-
-logging:
-  level:
-    org:
-      springframework:
-        data:
-          mongodb:
-            core:
-              MongoTemplate: DEBUG
