+++ web/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id 'war'
+}
+
+bootWar {
+    enabled = true
+    archiveName 'DRINK.WEB.war'
+}
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+}
